app/requirements/dev.txt
-r base.txt

django-debug-toolbar==4.4.6
django-coverage-plugin==3.1.0
factory-boy==3.3.0
pre-commit==3.7.1
pytest-asyncio==0.23.8
pytest-django==4.8.0
ruff==0.5.2
